move IResourceChange to API
